Phase 4 (13/n): export backtick (stateexporter), PekkoHttpServerComponents, PekkoSSLConfig, breakOut→toSet, Right/rbac/polyfill/cookies toSeq

+++ otoroshi/app/next/plugins/graphql.scala
@@ -677,7 +677,7 @@ class GraphQLBackend extends NgBackendCall {
             case d: Boolean       => JsBoolean(d)
             case Some(s: String)  => JsString(s)
             case Some(i: Int)     => JsNumber(i)
-            case Some(f: Float)   => JsNumber(f)
+            case Some(f: Float)   => JsNumber(f.toDouble)
             case Some(d: Boolean) => JsBoolean(d)
             case a                => JsString(String.valueOf(a))
           }))
